# src/main/kotlin/com/ruzzante/arquiteturahexagonal/domain/optional/Optional.kt
package com.ruzzante.arquiteturahexagonal.domain.optional

class Optional (
    val direcaoHidraulica: Boolean,
    val cambioAutomatico: Boolean,
    val vidroEletrico: Boolean,
    val travaEletrica: Boolean,
    val som: Boolean,
    val alarme: Boolean,
    val cameraDeRe: Boolean
    ){

}